% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.R
\name{print.pcc}
\alias{print.pcc}
\title{External functions}
\usage{
\method{print}{pcc}(x, ...)
}
\arguments{
\item{x}{A PCC or PRCC object}

\item{\dots}{Currently ignored}
}
\description{
This function is derived from package "sensitivity", please proceed to its help files.
}
