% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcv.R
\name{cv}
\alias{cv}
\title{Coefficient of Variation}
\usage{
cv(x, ...)
}
\arguments{
\item{x}{Any numeric vector (or other data type for which \code{sd} and \code{mean} methods are defined)}

\item{\dots}{Additional parameters for the \code{sd} and \code{mean} functions (such as \code{na.rm=TRUE})}
}
\description{
Returns the coefficient of variation of a sample.
}
