% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pic.R
\name{pic}
\alias{pic}
\alias{pic.LHS}
\alias{pic.default}
\alias{print.pic}
\title{Partial Inclination Coefficient
Estimates the partial inclination coefficient of a model response
in relation with all model input variables.}
\source{
based on code of pcc function on sensitivity package
 by Gilles Pujol, Bertrand Iooss, Alexandre Janon.
}
\usage{
pic(X, y, nboot, conf, ...)

\method{pic}{LHS}(X, y = NULL, nboot = 0, conf = 0.95, ...)

\method{pic}{default}(X, y, nboot = 0, conf = 0.95, ...)

\method{print}{pic}(x, ...)
}
\arguments{
\item{X}{A data frame (or object coercible by \code{as.data.frame})
containing the design of experiments (model input variables).}

\item{y}{A vector containing the responses corresponding to the design
of experiments (model output variables).}

\item{nboot}{Number of bootstrap replicates}

\item{conf}{Confidence of the bootstrap interval}

\item{\dots}{Currently ignored.}

\item{x}{The object returned by \code{pic}.}
}
\description{
Partial Inclination Coefficient
Estimates the partial inclination coefficient of a model response
in relation with all model input variables.
}
\author{
Andre Chalom, based on code by Gilles Pujol, Bertrand Iooss, Alexandre Janon
}
